requirements.txt
# CCTNS-GridX Dependencies
# Core Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9

# Security
PyJWT==2.9.0
bcrypt==4.2.0
cryptography==43.0.0

# AI / ML
scikit-learn==1.5.0
numpy==1.26.4
scipy==1.14.0
pandas==2.2.2

# Utilities
aiofiles==24.1.0
python-dateutil==2.9.0
